[PLAYER-5447] Refactor shared component to use own styles
