+++ sdk/react/src/shared/ItemSelectionScrollView/ItemSelectionList/ItemSelectionList.styles.js
@@ -1,0 +1,12 @@
+// @flow
+
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  column: {
+    //
+  },
+  row: {
+    //
+  },
+});
