+++ sdk/react/src/shared/RectangularButton/RectangularButton.js
@@ -2,8 +2,7 @@ import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 import { Text, TouchableHighlight, View } from 'react-native';
 
-// TODO(styles): Prefer own styles instead.
-import styles from '../styles/rectangularButton.styles';
+import styles from './RectangularButton.styles';
 
 export default class RectangularButton extends Component {
   static propTypes = {
