+++ sdk/react/src/shared/SkipButton/SkipButton.js
@@ -5,8 +5,7 @@ import { Animated, TouchableHighlight } from 'react-native';
 import { STRING_CONSTANTS } from '../../constants';
 import AccessibilityUtils from '../../lib/accessibility';
 
-// TODO(styles): Prefer own styles instead.
-import styles from '../styles/rectangularButton.styles';
+import styles from './SkipButton.styles';
 
 export default class SkipButton extends Component {
   static propTypes = {
