+++ sdk/react/src/shared/SkipButton/SkipButton.styles.js
@@ -1,0 +1,9 @@
+// @flow
+
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  buttonTextStyle: {
+    color: 'white',
+  },
+});
