+++ sdk/react/src/shared/VideoViewPlayPause/VideoViewPlayPause.js
@@ -8,8 +8,7 @@ import AccessibilityUtils from '../../lib/accessibility';
 import * as Utils from '../../lib/utils';
 import SkipButton from '../SkipButton';
 
-// TODO(styles): Prefer own styles instead.
-import styles from '../styles/rectangularButton.styles';
+import styles from './VideoViewPlayPause.styles';
 
 const PLAY = 'play';
 const PAUSE = 'pause';
