+++ sdk/react/src/shared/VideoViewPlayPause/VideoViewPlayPause.styles.js
@@ -1,0 +1,9 @@
+// @flow
+
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+  buttonTextStyle: {
+    color: 'white',
+  },
+});
